Add user-message relationship and integrate authentication for message endpoints

+++ frontend/src/components/chat/index.tsx
@@ -1,30 +1,49 @@
 import { useEffect, useState } from 'react';
 import { Message } from '@/types';
+import { useAuth } from '../../../context/AuthContext';
 
 function Chat(){
     const [messages, setMessages] = useState<Message[]>([])
     const [input, setInput] = useState<string>("")
+    const { user, token } = useAuth();
 
-    const fetchMessages = () => {
-        fetch("http://127.0.0.1:8000/api/messages")
-        .then(res => res.json())
-        .then(data => setMessages(data));
+    const fetchMessages = async () => {
+        if (!token) return
+        const res = await fetch("http://127.0.0.1:8000/api/messages/", {
+            headers: {
+                "Authorization" : `Bearer ${token}`
+            }
+        });
+        const data = await res.json();
+        setMessages(data);
     };
 
     const sendMessage = async () => {
         if (!input) return;
-        await fetch("http://127.0.0.1:8000/api/messages",{
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({content: input})
-        });
-        setInput("");
-        fetchMessages();
+        try{
+            await fetch("http://127.0.0.1:8000/api/messages/",{
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${token}`
+                },
+                body: JSON.stringify({content: input})
+            });
+            setInput("");
+            await fetchMessages();
+        } catch (err) {
+            console.error("Failed to send message:", err);
+        }
     }
 
     const deleteMessage = async (id: number) => {
+        if (!token) return;
+
         await fetch(`http://127.0.0.1:8000/api/messages/${id}`,{
             method: "DELETE",
+            headers: {
+                "Authorization": `Bearer ${token}`
+            }
         }); 
         await fetchMessages();
     }
@@ -47,7 +66,7 @@ function Chat(){
                 {messages.map((msg: Message) => (
                 <li key={msg.id}>
                     <button style={{border: "1px solid red"}} onClick={() => deleteMessage(msg.id)}>X</button>
-                    {msg.content}
+                    {msg.user && msg.user.name}: {msg.content}
                 </li>
                 ))}
             </ul>
